+++ common/src/main/java/earth/terrarium/cadmus/common/claims/maxclaims/CadmusMaxClaimProvider.java
@@ -3,7 +3,10 @@ package earth.terrarium.cadmus.common.claims.maxclaims;
 import earth.terrarium.cadmus.api.claims.maxclaims.MaxClaimProvider;
 import earth.terrarium.cadmus.common.util.ModGameRules;
 import net.minecraft.server.MinecraftServer;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.player.Player;
+
+import java.util.UUID;
 
 public class CadmusMaxClaimProvider implements MaxClaimProvider {
 
@@ -15,14 +18,23 @@ public class CadmusMaxClaimProvider implements MaxClaimProvider {
     public void removeTeam(String id, MinecraftServer server) {
     }
 
-    // The Cadmus one just uses the game rule for the max claims
     @Override
     public int getMaxClaims(String id, MinecraftServer server, Player player) {
-        return ModGameRules.getOrCreateIntGameRule(player.level(), ModGameRules.RULE_MAX_CLAIMED_CHUNKS);
+        return getMaxClaims(id, (ServerLevel) player.level(), player.getUUID());
+    }
+
+    @Override
+    public int getMaxClaims(String id, ServerLevel level, UUID player) {
+        return ModGameRules.getOrCreateIntGameRule(level, ModGameRules.RULE_MAX_CLAIMED_CHUNKS);
     }
 
     @Override
     public int getMaxChunkLoaded(String id, MinecraftServer server, Player player) {
-        return ModGameRules.getOrCreateIntGameRule(player.level(), ModGameRules.RULE_MAX_CHUNK_LOADED);
+        return getMaxChunkLoaded(id, (ServerLevel) player.level(), player.getUUID());
+    }
+
+    @Override
+    public int getMaxChunkLoaded(String id, ServerLevel level, UUID player) {
+        return ModGameRules.getOrCreateIntGameRule(level, ModGameRules.RULE_MAX_CHUNK_LOADED);
     }
 }