+++ common/src/main/java/earth/terrarium/cadmus/common/commands/claims/admin/AdminCommands.java
@@ -6,12 +6,12 @@ import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.suggestion.SuggestionProvider;
 import com.mojang.datafixers.util.Pair;
 import com.teamresourceful.resourcefullib.common.utils.CommonUtils;
+import earth.terrarium.cadmus.api.claims.ClaimApi;
 import earth.terrarium.cadmus.common.claims.ClaimHandler;
 import earth.terrarium.cadmus.common.claims.ClaimType;
 import earth.terrarium.cadmus.common.commands.claims.ClaimException;
 import earth.terrarium.cadmus.common.commands.claims.CommandHelper;
 import earth.terrarium.cadmus.common.teams.TeamHelper;
-import earth.terrarium.cadmus.common.util.ModUtils;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
 import net.minecraft.commands.SharedSuggestionProvider;
@@ -115,7 +115,7 @@ public class AdminCommands {
             boolean isMember = TeamHelper.isMember(claimData.getFirst(), player.server, player.getUUID());
             throw isMember ? ClaimException.ALREADY_CLAIMED_CHUNK : ClaimException.CHUNK_ALREADY_CLAIMED;
         }
-        ModUtils.claim(id, player.serverLevel(), pos, chunkloaded ? ClaimType.CHUNK_LOADED : ClaimType.CLAIMED);
+        ClaimApi.API.claim(player.serverLevel(), pos, id, chunkloaded);
         if (chunkloaded) {
             player.displayClientMessage(CommonUtils.serverTranslatable("text.cadmus.claiming.chunk_loaded_chunk_at", pos.x, pos.z), false);
         } else {
@@ -126,7 +126,7 @@ public class AdminCommands {
     public static void unclaim(ServerPlayer player, ChunkPos pos) throws ClaimException {
         Pair<String, ClaimType> claimData = ClaimHandler.getClaim(player.serverLevel(), pos);
         if (claimData == null) throw ClaimException.CHUNK_NOT_CLAIMED;
-        ModUtils.unclaim(claimData.getFirst(), player.serverLevel(), pos);
+        ClaimApi.API.unclaim(player.serverLevel(), pos, claimData.getFirst());
         player.displayClientMessage(CommonUtils.serverTranslatable("text.cadmus.unclaiming.unclaimed_chunk_at", pos.x, pos.z), false);
     }
 
@@ -134,7 +134,7 @@ public class AdminCommands {
         Pair<String, ClaimType> claimData = ClaimHandler.getClaim(player.serverLevel(), pos);
         if (claimData == null) throw ClaimException.CHUNK_NOT_CLAIMED;
         if (!claimData.getFirst().equals(id)) throw ClaimException.DONT_OWN_CHUNK;
-        ModUtils.unclaim(id, player.serverLevel(), pos);
+        ClaimApi.API.unclaim(player.serverLevel(), pos, id);
         player.displayClientMessage(CommonUtils.serverTranslatable("text.cadmus.unclaiming.unclaimed_chunk_at", pos.x, pos.z), false);
     }